Made it so unless the user types a new password on login fail the login action will not execute
This is ideally to make it so people cannot spam the login field. Ideally we do not want to send a password we know that already causes a failed login. Password must be changed by a keystroke. The use of more events helps prevent abuse of this system since these things as far as I know cannot be reliably simulated