Added comments. Added in an onload function for the shipping fields after the garlic retrieves. Added in login pull back.
Added in an onload function for the shipping fields after the garlic retrieves. Pipes all shipping values to the shipping details spot. This takes care of any renegade data that did not want to persist

Login Pullback: If user shows up on anything other than the customer info tab and isnt logged in but we found an account for the persisted email, pull them back to customer info and focus on the password field.