Updated composer. Consequently updated WordPress simple settings. Fixed error being generated from Settings Manager in main.
Updated WordPress_SimpleSettings to 0.4
Main was generating an error where the simple settings was being used because $this in a static context has no reference. So snagged the main instance and used that to point to the objects required instead. Helps keep errors only popping up where there are actual errors.