Fixed an issue where a failing response from stripe on the woocommerce backend call was preventing subsequent stripe calls with the same data.
This was resulting because the stripe call, while successfull really was a bad data response. However the way the system works, we don't find that out until WC does more backend validation of the token. The resolution to this bug was to make sure the stripe_token field (hidden field) was removed on response failure from woocommerce.