Had to partially refactor the way compatibility classes were down to allow for earlier setup in the main setup method scheme. Now compatibility classes can choose when (before/after) setup when to register their compatibility class.
Also added event related functions in Main. Moved all the events not relating to entry into Main for now. Will make an event class setup later.

If klarna is detected to be active but not running everything acts as normal. However if the klarna button has been clicked the breadcrumb does not show and EasyTabs does not initialize.