Use new ActivationManager instead of booster seat Activator.
We needed more flexibility here, but rather than refactor BoosterSeat, I built a new ActivationManager class that allows just about any evaluation to be performed.

The weakness of the regular Activator is that it assumes that plugins are always in the expected directories, and it also doesn't handle detection of network activated plugins.