Change values of ship_to_different_address to 1 and 2 from 0 and 1
WooCommerce natively interprets ship_to_different_address as empty or not empty. Because we are using this flag in the opposite manner that it actually means AND because we will always send both addresses, we need to send non-empty values.

Previous commit contains part of this work in TabContainer.ts.