Parsley refactor.
Because Parsley magic validation likes to muck up things, we are completely disabling it by adding formnovalidate to our form tags.

Then, we add a separate billing field validation that only runs when the cart needs shipping. If the billing fields don't validate, it prevents the complete order from running.

If they do, it allows it to continue.

We also removed some legacy references to validateZip which are no longer needed.